Correctly use V2 API namespace
