--- app/views/api/discovers/create.json.rabl
@@ -1,3 +1,0 @@
-object @discovered_host
-
-extends "api/discovers/show"
