--- app/views/api/discovers/index.json.rabl
@@ -1,3 +1,0 @@
-collection @discovers
-
-extends "api/discovers/main"
