+++ app/views/api/v2/discovers/base.json.rabl
