+++ app/views/api/v2/discovers/create.json.rabl
@@ -1,0 +1,3 @@
+object @discovered_host
+
+extends "api/v2/discovers/show"
