+++ app/views/api/v2/discovers/index.json.rabl
@@ -1,0 +1,3 @@
+collection @discovers
+
+extends "api/v2/discovers/main"
